Return whether deprovisioning has completed

+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/maintenance/HostDeprovisioner.java
@@ -46,8 +46,8 @@ public class HostDeprovisioner extends NodeRepositoryMaintainer {
                 //   holding it over longer periods
                 // * We are about to remove these hosts anyway, so only reason we'd want to hold the lock is
                 //   if we want to support aborting deprovision if operator manually intervenes
-                hostProvisioner.deprovision(host);
-                nodeRepository().nodes().removeRecursively(host, true);
+                if (hostProvisioner.deprovision(host))
+                    nodeRepository().nodes().removeRecursively(host, true);
             } catch (RuntimeException e) {
                 failures++;
                 log.log(Level.WARNING, "Failed to deprovision " + host.hostname() + ", will retry in " + interval(), e);
